Validação empresa_id no insert, update e delete

+++ app/Http/Controllers/CardapioController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Cardapio;
+use App\Models\Empresa;
 use App\Repositories\CardapioRepository;
 use Illuminate\Http\Request;
 use function GuzzleHttp\Promise\all;
@@ -75,6 +76,15 @@ class CardapioController extends Controller
      */
     public function store(Request $request)
     {
+        $empresa = Empresa::find($request->empresa_id);
+        if ($empresa === null) {
+            return response()->json(['erro' => 'Usuário não tem permissão de inserir o recurso solicitado!'], 403);
+        }
+
+        if ($empresa->user_id != auth()->user()->id) {
+            return response()->json(['erro' => 'Usuário não tem permissão de inserir o recurso solicitado!'], 403);
+        }
+
         // Recebe a request e valida os campos
         $request->validate($this->cardapio->rules());        
         // Salva a request na tabela e retorna o registro inserido
@@ -118,6 +128,10 @@ class CardapioController extends Controller
             return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe!'], 404);
         }
 
+        if ($cardapio->user_id != auth()->user()->id) {
+            return response()->json(['erro' => 'Usuário não tem permissão de alterar o recurso solicitado!'], 403);
+        }
+
         if ($request->method() === 'PATCH') {
             $regrasDinamicas = [];
             // Percorre todas as regras definidas no model
@@ -157,6 +171,11 @@ class CardapioController extends Controller
         if ($cardapio === null) {
             return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe!'], 404);
         }
+
+        if ($cardapio->user_id != auth()->user()->id) {
+            return response()->json(['erro' => 'Usuário não tem permissão de deletar o recurso solicitado!'], 403);
+        }
+
         // Deleta o registro selecionado
         $cardapio->delete();
